requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.0
MarkupSafe==1.1.1
nltk==3.5
numpy==1.20.0
oauthlib==3.1.0
pandas==1.1.5
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
textblob==0.15.3
tqdm==4.56.0
tweepy==3.10.0
urllib3==1.26.3
Werkzeug==1.0.1
XlsxWriter==1.3.7
